data: Rename undo checkpoint module
Undo stack orchestration no longer sits in a generic data.undo module. The
neighboring undo_refs, undo_restore, and undo_worktree modules already name
their narrower responsibilities.

Callers use the module to create checkpoints and move through the undo and
redo stacks. Naming it undo_checkpoints makes that contract visible at
import sites.

This commit renames data.undo to data.undo_checkpoints, updates command
imports, and refreshes the module docstring.

Undo checkpoint orchestration now sits beside its ref, restore, and
worktree collaborators under matching names.
