+++ src/git_stage_batch/commands/batch_source/apply_action.py
@@ -22,7 +22,7 @@ from ...batch.submodule_pointer import (
     is_batch_submodule_pointer,
 )
 from ...data.session import snapshot_file_if_untracked
-from ...data.undo import undo_checkpoint
+from ...data.undo_checkpoints import undo_checkpoint
 from ...exceptions import (
     AtomicUnitError,
     CommandError,
