+++ src/git_stage_batch/data/undo_checkpoints.py
@@ -1,4 +1,4 @@
-"""Undo checkpoint orchestration."""
+"""Undo checkpoint stack orchestration."""
 
 from __future__ import annotations
